+++ fideslog/api/database/data_access.py
@@ -9,8 +9,6 @@ from sqlalchemy.exc import DBAPIError
 
 from fideslog.api.models.analytics_event import AnalyticsEvent
 
-# from fideslog.api.database.database import get_db
-
 from fideslog.api.database.models import AnalyticsEvent as AnalyticsEventORM
 from fideslog.api.database.models import APIKey
